Increase functionality of Projects, start on mbl
Read More buttons now only show if the page has a body associated with
it in its .md file or a specific url in its front matter.

Photos for the ProjectCard are now defined in the page.md frontmatter,
as an `image:`.

Refactoring SCSS for mobile-first, waiting for @bohuim's changes to
continue w mobile.